etl/requirements.txt
p4d
psycopg2-binary
schedule
python-dotenv
ruff
pytest
